+++ packages/components-angular/src/app/design-tokens/design-tokens-spacing-example.component.ts
@@ -6,12 +6,13 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
   template: `
     <div>
       <div class="wrapper">
-        <h3 class="heading">Spacing Fluid (XS - XL)</h3>
+        <h3 class="heading">Spacing Fluid (XS - XXL)</h3>
         <div class="spacing-fluid-x-small tile"></div>
         <div class="spacing-fluid-small tile"></div>
         <div class="spacing-fluid-medium tile"></div>
         <div class="spacing-fluid-large tile"></div>
         <div class="spacing-fluid-x-large tile"></div>
+        <div class="spacing-fluid-xx-large tile"></div>
       </div>
       <div class="wrapper">
         <h3 class="heading">Spacing Static (XS - XXL)</h3>