+++ packages/components-angular/src/app/pages/overview.component.ts
@@ -51,11 +51,11 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
         </div>
 
         <div class="playground light" title="should render default headline">
-          <p-headline>The quick brown fox jumps over the lazy dog</p-headline>
+          <p-headline>Some Headline</p-headline>
         </div>
 
         <div class="playground light" title="should render default display">
-          <p-display>The quick brown fox jumps over the lazy dog</p-display>
+          <p-display>Some Display</p-display>
         </div>
 
         <div class="playground light" title="should render default banner">
@@ -164,6 +164,10 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 
         <div class="playground light" title="should render default marque">
           <p-marque></p-marque>
+        </div>
+
+        <div class="playground light" title="should render default model-signature">
+          <p-model-signature></p-model-signature>
         </div>
 
         <div class="playground light" title="should render default pagination">
@@ -308,11 +312,11 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
         </div>
 
         <div class="playground light" title="should render default headline with custom prefix">
-          <my-prefix-p-headline p-headline>The quick brown fox jumps over the lazy dog</my-prefix-p-headline>
+          <my-prefix-p-headline p-headline>Some Headline</my-prefix-p-headline>
         </div>
 
         <div class="playground light" title="should render default display with custom prefix">
-          <my-prefix-p-display p-display>The quick brown fox jumps over the lazy dog</my-prefix-p-display>
+          <my-prefix-p-display p-display>Some Display</my-prefix-p-display>
         </div>
 
         <div class="playground light" title="should render default banner with custom prefix">
@@ -424,6 +428,10 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 
         <div class="playground light" title="should render default marque with custom prefix">
           <my-prefix-p-marque p-marque></my-prefix-p-marque>
+        </div>
+
+        <div class="playground light" title="should render default model-signature with custom prefix">
+          <my-prefix-p-model-signature p-model-signature></my-prefix-p-model-signature>
         </div>
 
         <div class="playground light" title="should render default pagination with custom prefix">