+++ packages/components-js/projects/partials/scripts/generateIconLinksPartial.ts
@@ -1,4 +1,4 @@
-import { minifyHTML, withoutTagsOption } from './utils';
+import { minifyHTML } from './utils';
 import { CDN_BASE_PATH_ICONS } from '../../../../../cdn.config';
 import { ICON_NAMES, ICONS_MANIFEST } from '@porsche-design-system/icons';
 
@@ -9,23 +9,17 @@ export const generateIconLinksPartial = (): string => {
 type GetIconLinksOptions = {
   icons?: IconName[];
   cdn?: Cdn;
-  ${withoutTagsOption}
   format?: Format;
-};
-type GetIconLinksOptionsFormatHtml = Omit<GetIconLinksOptions, 'withoutTags'> & { format: 'html' };
-type GetIconLinksOptionsFormatJsx = Omit<GetIconLinksOptions, 'withoutTags'> & { format: 'jsx' };
-type GetIconLinksOptionsWithoutTags = Omit<GetIconLinksOptions, 'format'>;`;
+};`;
 
   const link = minifyHTML('<link rel="prefetch" href="${url}" as="image" type="image/svg+xml" crossorigin>');
 
-  const func = `export function getIconLinks(opts?: GetIconLinksOptionsFormatJsx): JSX.Element;
-export function getIconLinks(opts?: GetIconLinksOptionsFormatHtml): string;
-export function getIconLinks(opts?: GetIconLinksOptionsWithoutTags): string[];
-export function getIconLinks(opts?: GetIconLinksOptions): string | string[] | JSX.Element {
-  const { icons, cdn, withoutTags, format }: GetIconLinksOptions = {
+  const func = `export function getIconLinks(opts?: GetIconLinksOptions & { format: 'jsx' }): JSX.Element;
+export function getIconLinks(opts?: GetIconLinksOptions): string;
+export function getIconLinks(opts?: GetIconLinksOptions): string | JSX.Element {
+  const { icons, cdn, format }: GetIconLinksOptions = {
     icons: ['arrow-right'],
     cdn: 'auto',
-    withoutTags: false,
     format: 'html',
     ...opts,
   };
@@ -49,10 +43,11 @@ Please use only valid icon names:
   const linksHtml = urls
     .map((url) => \`${link}\`)
     .join('');
-  const linksJsx = urls.map((url, index) => <link key={index} rel="prefetch" href={url} as="image" type="image/svg+xml" crossOrigin="true" />);
+  const linksJsx = urls.map((url, index) => <link key={index} rel="prefetch" href={url} as="image" type="image/svg+xml" crossOrigin="" />);
 
-  const markup = format === 'html' ? linksHtml : <>{linksJsx}</>;
-  return withoutTags ? urls : markup;
+  return format === 'html'
+    ? linksHtml
+    : <>{linksJsx}</>;
 }`;
 
   return [types, func].join('\n\n');