+++ packages/components-js/scripts/convertToNextJsVRTPage.ts
@@ -1,58 +1,16 @@
-import { convertToReactVRTPage } from './convertToReactVRTPage';
 import * as fs from 'fs';
 import * as path from 'path';
 import { paramCase } from 'change-case';
-
-type Characteristics = {
-  usesSetAllReady: boolean;
-  usesComponentsReady: boolean;
-  usesToast: boolean;
-  isIconPage: boolean;
-  usesQuerySelector: boolean;
-  usesPrefixing: boolean;
-  isOverviewPage: boolean;
-};
+import { convertToReactVRTPage, ReactCharacteristics } from './convertToReactVRTPage';
 
 const sourceBasePath = path.resolve(__dirname, '../../components-react/src');
 const pollComponentsReadyFilePath = path.resolve(sourceBasePath, 'pollComponentsReady.ts');
 const pollComponentsReadyFileContent = fs.readFileSync(pollComponentsReadyFilePath, 'utf8');
 
 export const convertToNextJsVRTPage = (
-  fileName: string,
-  fileContent: string,
-  template: string,
-  style: string,
-  script: string,
-  toastText: string,
-  characteristics: Characteristics
+  ...params: [string, string, string, string, string, string, ReactCharacteristics]
 ): { fileName: string; fileContent: string } => {
-  const {
-    usesSetAllReady,
-    usesComponentsReady,
-    usesToast,
-    isIconPage,
-    usesQuerySelector,
-    usesPrefixing,
-    isOverviewPage,
-  } = characteristics;
-
-  const { fileName: convertedFileName, fileContent: convertedFileContent } = convertToReactVRTPage(
-    fileName,
-    fileContent,
-    template,
-    style,
-    script,
-    toastText,
-    {
-      usesSetAllReady,
-      usesComponentsReady,
-      usesToast,
-      isIconPage,
-      usesQuerySelector,
-      usesPrefixing,
-      isOverviewPage,
-    }
-  );
+  const { fileName: convertedFileName, fileContent: convertedFileContent } = convertToReactVRTPage(...params);
 
   let newFileContent = convertedFileContent
     .replace(/\/\* Auto Generated File \*\//, "$&\nimport type { NextPage } from 'next';")