+++ packages/components-js/tests/e2e/puppeteer/specs/stepper-horizontal.e2e.ts
@@ -127,7 +127,7 @@ describe('validation', () => {
   });
 });
 
-xdescribe('scrolling', () => {
+describe('scrolling', () => {
   it('should scroll current step into view', async () => {
     await initStepperHorizontal({ amount: 9, currentStep: 3, isWrapped: true });
     const [, , , step4] = await getStepItems();
@@ -288,7 +288,7 @@ describe('events', () => {
 
     await addEventListener(host, 'stepChange');
 
-    await page.mouse.click((await getOffsetWidth(item1)) + 8, 18);
+    await page.mouse.click((await getOffsetWidth(item1)) + 6, 18);
     expect((await getEventSummary(host, 'stepChange')).counter).toBe(0);
 
     await item1.click();