+++ packages/components-react/projects/nextjs/pages/overview.tsx
@@ -1,7 +1,7 @@
 /* Auto Generated File */
 import Image from 'next/image';
 import type { NextPage } from 'next';
-import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PCarousel, PCheckboxWrapper, PContentWrapper, PDisplay, PDivider, PFieldsetWrapper, PFlex, PFlexItem, PGrid, PGridItem, PHeading, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PLinkTile, PMarque, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PScroller, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PStepperHorizontal, PStepperHorizontalItem, PSwitch, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react/ssr';
+import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PCarousel, PCheckboxWrapper, PContentWrapper, PDisplay, PDivider, PFieldsetWrapper, PFlex, PFlexItem, PGrid, PGridItem, PHeading, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PLinkTile, PMarque, PModelSignature, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PScroller, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PStepperHorizontal, PStepperHorizontalItem, PSwitch, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react/ssr';
 
 const OverviewPage: NextPage = (): JSX.Element => {
   const style = `
@@ -54,11 +54,11 @@ const OverviewPage: NextPage = (): JSX.Element => {
           </div>
 
           <div className="playground light" title="should render default headline">
-            <PHeadline>The quick brown fox jumps over the lazy dog</PHeadline>
+            <PHeadline>Some Headline</PHeadline>
           </div>
 
           <div className="playground light" title="should render default display">
-            <PDisplay>The quick brown fox jumps over the lazy dog</PDisplay>
+            <PDisplay>Some Display</PDisplay>
           </div>
 
           <div className="playground light" title="should render default banner">
@@ -167,6 +167,10 @@ const OverviewPage: NextPage = (): JSX.Element => {
 
           <div className="playground light" title="should render default marque">
             <PMarque />
+          </div>
+
+          <div className="playground light" title="should render default model-signature">
+            <PModelSignature />
           </div>
 
           <div className="playground light" title="should render default pagination">
@@ -312,11 +316,11 @@ const OverviewPage: NextPage = (): JSX.Element => {
             </div>
 
             <div className="playground light" title="should render default headline with custom prefix">
-              <PHeadline>The quick brown fox jumps over the lazy dog</PHeadline>
+              <PHeadline>Some Headline</PHeadline>
             </div>
 
             <div className="playground light" title="should render default display with custom prefix">
-              <PDisplay>The quick brown fox jumps over the lazy dog</PDisplay>
+              <PDisplay>Some Display</PDisplay>
             </div>
 
             <div className="playground light" title="should render default banner with custom prefix">
@@ -428,6 +432,10 @@ const OverviewPage: NextPage = (): JSX.Element => {
 
             <div className="playground light" title="should render default marque with custom prefix">
               <PMarque />
+            </div>
+
+            <div className="playground light" title="should render default model-signature with custom prefix">
+              <PModelSignature />
             </div>
 
             <div className="playground light" title="should render default pagination with custom prefix">