+++ packages/storefront/storefront.config.ts
@@ -42,10 +42,24 @@ export const config: StorefrontConfig = {
     'Next Js': {
       // @ts-ignore
       'Getting Started': [() => import('@/pages/start-coding/next-js/getting-started.md')],
-      // @ts-ignore
-      'Ssr Support': [() => import('@/pages/start-coding/next-js/ssr-support.md')],
+      'Ssr Support': [
+        // @ts-ignore
+        () => import('@/pages/start-coding/next-js/ssr-support.md'),
+        // @ts-ignore
+        () => import('@/pages/start-coding/ssr-support-table.md'),
+      ],
       // @ts-ignore
       Testing: [() => import('@/pages/start-coding/next-js/testing.md')],
+    },
+    Remix: {
+      // @ts-ignore
+      'Getting Started': [() => import('@/pages/start-coding/remix/getting-started.md')],
+      'Ssr Support': [
+        // @ts-ignore
+        () => import('@/pages/start-coding/remix/ssr-support.md'),
+        // @ts-ignore
+        () => import('@/pages/start-coding/ssr-support-table.md'),
+      ],
     },
   },
   News: {
@@ -247,6 +261,14 @@ export const config: StorefrontConfig = {
       // @ts-ignore
       Props: [() => import('@/../../components/src/components/modal/modal.props.md')],
     },
+    'Model Signature': {
+      // @ts-ignore
+      Examples: [() => import('@/../../components/src/components/model-signature/model-signature.examples.md')],
+      // @ts-ignore
+      Usage: [() => import('@/../../components/src/components/model-signature/model-signature.usage.md')],
+      // @ts-ignore
+      Props: [() => import('@/../../components/src/components/model-signature/model-signature.props.md')],
+    },
     Notifications: {
       // @ts-ignore
       Introduction: [() => import('@/pages/components/notifications/notifications.usage.md')],
@@ -255,7 +277,7 @@ export const config: StorefrontConfig = {
         () => import('@/../../components/src/components/inline-notification/inline-notification.examples.md'),
       ],
       // @ts-ignore
-      Toast: [() => import('@/../../components/src/components/toast/toast/toast.examples.md')],
+      Toast: [() => import('@/../../components/src/components/toast/toast.examples.md')],
       // @ts-ignore
       Banner: [() => import('@/../../components/src/components/banner/banner.examples.md')],
       // @ts-ignore