+++ packages/utilities/projects/utilities/src/js/spacing/spacingFluidLarge.ts
@@ -1,1 +1,1 @@
-export const spacingFluidLarge = 'clamp(3.5rem, 2.22vw + 3.06rem, 5.5rem)';
+export const spacingFluidLarge = 'clamp(32px, 2.75vw + 23px, 76px)';