+++ packages/utilities/projects/utilities/src/js/spacing/spacingFluidXLarge.ts
@@ -1,1 +1,1 @@
-export const spacingFluidXLarge = 'clamp(4rem, 10vw + 2rem, 13rem)';
+export const spacingFluidXLarge = 'clamp(48px, 3vw + 38px, 96px)';