+++ packages/components/CHANGELOG.md
@@ -9,9 +9,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Changed
+
+- Background color of `Scroller`'s `prev` and `next` buttons in dark theme
+- Partials: Removed deprecated `withoutTags` option for all partials, please use `format: 'jsx'` instead
+- `Divider` and `Button Group` match new design language.
+- `Text List` matches new design language
+
 #### Added
 
+- `Model Signature`
 - Vue: plugin functions `createPorscheDesignSystem` and `usePorscheDesignSystemPlugin`
+
+### [3.0.0-alpha.1] - 2023-02-16
+
+#### Added
+
+- Porsche Next font supports Vietnamese charset
+- Prop `color` of `Icon` supports `disabled`
+- React: `patchRemixRunProcessBrowserGlobalIdentifier` binary to support SSR components with Remix
+
+#### Changed
+
+- `Stepper Horizontal` matches new design language
+- Styles: Optimize design tokens "spacing", "typography" and "theme" provided by styles sub-package
+  `@porsche-design-system/components-{js|angular|react|vue}/styles`
+- Styles: Use calc() instead of max() to calculate padding for `gridStyle` (JS) and `pds-grid` (SCSS)
+- Styles: `gridStyle` (JS) and `pds-grid` (SCSS) uses optimized grid gap
 
 ### [3.0.0-alpha.0] - 2023-02-08
 
@@ -383,6 +407,21 @@ to the new values since those ones will be removed with next major version.
 + …addMessage({ text: `Some message`, state: 'info' })
 ```
 
+##### Divider:
+
+- Prop values `neutral-contrast-low | neutral-contrast-medium | neutral-contrast-high` of `color` prop are deprecated.
+
+```diff
+- <p-divider color="neutral-contrast-low"></p-divider>
++ <p-divider color="contrast-low"></p-divider>
+
+- <p-divider color="neutral-contrast-medium"></p-divider>
++ <p-divider color="contrast-medium"></p-divider>
+
+- <p-divider color="neutral-contrast-high"></p-divider>
++ <p-divider color="contrast-high"></p-divider>
+```
+
 #### Added
 
 - `Display` component
@@ -399,10 +438,10 @@ to the new values since those ones will be removed with next major version.
   `Radio Button Wrapper`, `Popover`, `Modal`, `Select Wrapper`, `Tag`, `Tag Dismissible`, `Textarea Wrapper`,
   `Inline Notification`, `Banner`, `Toast`, `Grid`, `Flex`, `Pagination`, `Scroller`, `Accordion`, `Text`,
   `Text Field Wrapper`, `Content Wrapper`, `Segmented Control`, `Tabs`, `Tabs Bar`, `Headline` and `Fieldset Wrapper`
-  matches new design language.
+  match new design language
 - `Icon` supports
   `primary | contrast-low | contrast-medium | contrast-high | notification-success | notification-warning | notification-error | notification-info | inherit`
-  for `color` prop.
+  for `color` prop
 - Default value of prop `width` of `Banner` has changed from `basic` to `extended`
 - Default value of prop `action-icon` of `Inline Notification` has changed from `arrow-head-right` to `arrow-right`
 - Default value of prop `name` of `Icon` has changed from `arrow-head-right` to `arrow-right`